Extract errors ui to a new module
