+++ collect_app/build.gradle
@@ -253,6 +253,7 @@ dependencies {
     implementation project(':projects')
     implementation project(':location')
     implementation project(':geo')
+    implementation project(':errors')
 
     implementation Dependencies.androidx_legacy_support
     implementation Dependencies.androidx_core_ktx
