+++ collect_app/src/main/AndroidManifest.xml
@@ -118,7 +118,9 @@ the specific language governing permissions and limitations under the License.
             android:name=".activities.FillBlankFormActivity"
             android:launchMode="singleTop" />
         <activity android:name=".activities.FormDownloadListActivity" />
-        <activity android:name="org.odk.collect.android.errors.ErrorActivity" />
+        <activity
+            android:name="org.odk.collect.errors.ErrorActivity"
+            android:theme="@style/Theme.Collect" />
         <activity
             android:name=".activities.DeleteSavedFormActivity"
             android:configChanges="orientation|screenSize" />
