+++ errors/src/main/java/org/odk/collect/errors/ErrorActivity.kt
@@ -1,10 +1,9 @@
-package org.odk.collect.android.errors
+package org.odk.collect.errors
 
 import android.os.Bundle
 import androidx.appcompat.widget.Toolbar
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import org.odk.collect.android.R
 import org.odk.collect.androidshared.utils.AppBarUtils.setupAppBarLayout
 import org.odk.collect.strings.localization.LocalizedActivity
 import org.odk.collect.strings.localization.getLocalizedString
