+++ errors/src/main/res/layout/activity_error.xml
