+++ settings.gradle
@@ -15,4 +15,5 @@ include ':material'
 include ':fragmentstest'
 include ':location'
 include ':geo'
+include ':errors'
 include ':externalapp'
